requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
importlib-metadata==3.4.0
mccabe==0.6.1
psycopg2==2.7.4
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
yapf==0.30.0
zipp==3.4.0
